auth_handler: always request consent on upstream OIDC IDP code flow
This change updates the code flow we perform with the upstream OIDC
IDP to always request consent via the prompt=consent query parameter.
This is required for us to receive a refresh token on each login
(i.e. for our access_type=offline parameter to always be honored
instead of if being honored on the first login and ignored for all
logins after that).

https://openid.net/specs/openid-connect-core-1_0.html#OfflineAccess

Signed-off-by: Monis Khan <mok@vmware.com>